+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Change
 
 - Now `tester.Detail.Error` includes `build_stderr`.
+- Now `tester.NewTestUnit` takes `*tester.Config` as argument. You have to create a `*tester.Config` instance before `tester.NewTestUnit`
 
 ## [0.5.1] - 2021-02-02